refactor: Webpack config, reduce dupe config
Upgrade webpack to v4.29
Remove react-router-config
Move bin folder to src
Remove logs folder
